Refactored CITRINE_SCOPE into DataConcepts
