+++ src/citrine/resources/material_run.py
@@ -8,11 +8,10 @@ from uuid import UUID
 import deprecation
 
 from citrine._rest.resource import Resource
-from citrine._serialization.serializable import CITRINE_SCOPE
 from citrine._serialization.properties import List as PropertyList
 from citrine._serialization.properties import Optional as PropertyOptional
 from citrine._serialization.properties import String, LinkOrElse, Mapping, Object
-from citrine.resources.data_concepts import DataConcepts
+from citrine.resources.data_concepts import DataConcepts, CITRINE_SCOPE
 from citrine.resources.material_spec import MaterialSpecCollection
 from citrine.resources.object_runs import ObjectRun, ObjectRunCollection
 from gemd.entity.file_link import FileLink
